Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/iframe-inheritance-expected.txt

Referrer Policy: iframes correctly inherit the ancestor's referrer


FAIL iframes correctly inherit the ancestor's referrer assert_equals: expected (string) "http://localhost:8800/" but got (undefined) undefined
